obs-ffmpeg: Create a new mpegts muxer output.
Currently the ffmpeg_mpegts_muxer output is integrated with ffmpeg-mux.
Both use obs native encoders in contrast with obs-ffmpeg-output which
relies on avcodec library.
This allowed easy implementation of srt, rist & hls protocols through
avformat library.
The main drawback is that obs-ffmpeg-mux exe doesn't allow for easy
debugging nor logging.
It was written initially as a separate binary designed for recording so
that if obs fails for some reason, the recording can still terminate
gracefully.
This commit is the first step towards a better support of srt & rist
protocols.
The ffmpeg_mpegts_muxer has been rewritten so that a pipe to the
ffmpeg-mux binary is not used any more.
The muxing to mpegts can be traced more easily in all its steps, except
 for the protocol handling which is still done by avformat.
Note that the mpegts muxer supports all protocols compatible with
mpegts, if supported by avformat (so udp, tcp, rtp ...).

Signed-off-by: pkv <pkv@obsproject.com>